requirements.txt
requests
python-dotenv
jupyter
pandas
plotly
numpy
